requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
redis>=5.0.0
django-redis>=5.4.0
gunicorn>=21.2.0
whitenoise>=6.6.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
Pillow>=10.1.0
reportlab>=4.0.7
xhtml2pdf>=0.2.11
